+++ cargo-nextest/src/lib.rs
@@ -8,8 +8,8 @@
 //! # Installation
 //!
 //! While you can install cargo-nextest from source, using the pre-built binaries is recommended.
-//! See [Pre-built binaries](https://nexte.st/book/pre-built-binaries) on the nextest site
-//! for more information.
+//! See [Pre-built binaries](https://nexte.st/docs/installation/pre-built-binaries) on the nextest
+//! site for more information.
 
 #![warn(missing_docs)]